Tao lop doi tuong OpflowConnector cho tung nhom ket noi giua Commander - Serverlet; Ho tro nhieu khai bao API
